updating and cleaning code

+++ inventory/src/main/java/com/riwi/MealMapSpring/Entities/DishesIngredients.java
@@ -26,5 +26,5 @@ public class DishesIngredients {
     private Ingredients ingredients;
 
     @Column(name = "quantity")
-    private int quantity;
+    private double quantity;
 }
